+++ CHANGELOG.md
@@ -9,15 +9,23 @@ This project adheres to [Semantic Versioning](http://semver.org/). Additionally,
 - **Fixed** for any bug fixes.
 - **Security** for any security changes or fixes for vulnerabilities.
 
-### **[0.0.3]**
+### **[0.0.4]**
 
 #### Added
 
+<<<<<<< HEAD
+=======
+### **[0.0.3] - 2017-08-31 [RELEASED]**
+
+#### Added
+
+>>>>>>> dev
 *  Support for kubernetes service annotations. [[BITE-1511](https://agile-jira.pearson.com/browse/BITE-1511)]
 *  Support for configuring horizontal pod autoscaling. [[BITE-1433](https://agile-jira.pearson.com/browse/BITE-1433)]
-*  Support for configuring container CPU resource requests. [[BITE-1433](https://agile-jira.pearson.com/browse/BITE-1433)]
+*  Added new environment operator endpoint for Pod Status. [[BITE-1484](https://agile-jira.pearson.com/browse/BITE-1484)]
 *  Custom Docker registry support added for pod spec. [[BITE-1448](https://agile-jira.pearson.com/browse/BITE-1448)]
 *  Environment Operator build/release pipeline now managed by TravisCI. [[BITE-1473](https://agile-jira.pearson.com/browse/BITE-1473)]
+*  Add error handling for secrets defined in environment.bitesize files for deployments. [[BITE-1465](https://agile-jira.pearson.com/browse/BITE-1465)]
 
 ### **[0.0.2] - 2017-01-17 [RELEASED]**